Refactor bot into package modules
